+++ email/src/folder/add/imap.rs
@@ -28,11 +28,12 @@ impl AddFolderImap {
 #[async_trait]
 impl AddFolder for AddFolderImap {
     async fn add_folder(&self, folder: &str) -> Result<()> {
-        info!("creating imap folder {folder}");
+        info!("imap: creating folder {folder}");
 
         let mut session = self.session.lock().await;
+        let config = &session.account_config;
 
-        let folder = session.account_config.get_folder_alias(folder);
+        let folder = config.get_folder_alias(folder);
         let folder_encoded = encode_utf7(folder.clone());
         debug!("utf7 encoded folder: {folder_encoded}");